Fixed Signup page as per PR review
