+++ src/components/SignupContent/Components/SignupContentList/SignupContentList.d.tsx
@@ -1,0 +1,5 @@
+import { TProposition } from "@/models/SignupProposition";
+
+export interface ISignupContentList {
+  signupProposition: TProposition;
+}
