+++ src/components/SignupContent/SignupContent.tsx
@@ -2,10 +2,10 @@ import { useTranslation } from "react-i18next";
 
 import { LanguageCode } from "@/models/enums/LanguageCode";
 
-import signuppropositions from "@/data/signup.json";
-import { SignupProposition } from "@/models/SignupProposition";
+import signupPropositionData from "@/data/signupPropositions.json";
+import { TProposition } from "@/models/SignupProposition";
 
-import SignupContentList from "../SignupContentList/SignupContentList";
+import SignupContentList from "./Components/SignupContentList/SignupContentList";
 
 const SignupContent = () => {
   const {
@@ -14,23 +14,22 @@ const SignupContent = () => {
   } = useTranslation();
 
   const currentLanguage = language as LanguageCode;
+  const propositionData: TProposition[] =
+    signupPropositionData[currentLanguage];
 
   return (
     <section className="pb-12 pl-8 pt-[137px] text-primary">
-      <h1 className="w-[430px] font-segoe text-4xl font-semibold">
+      <h1 className="max-w-[430px] font-segoe text-4xl font-semibold">
         {t("pages.signup.signupContent.title")}
       </h1>
 
       <ul className="mt-[83px] flex flex-col gap-y-12">
-        {(signuppropositions[currentLanguage] as SignupProposition[]).map(
-          (signupProposition) => (
-            <SignupContentList
-              key={signupProposition.id}
-              signupProposition={signupProposition}
-              currentLanguage={currentLanguage}
-            />
-          ),
-        )}
+        {propositionData.map((signupProposition: TProposition) => (
+          <SignupContentList
+            key={`signProposition-${signupProposition.id}`}
+            signupProposition={signupProposition}
+          />
+        ))}
       </ul>
     </section>
   );
