--- src/components/SignupContentList/SignupContentList.d.tsx
@@ -1,7 +1,0 @@
-import { SignupProposition } from "@/models/SignupProposition";
-import { LanguageCode } from "@/models/enums/LanguageCode";
-
-export interface ISignupContentList {
-  signupProposition: SignupProposition;
-  currentLanguage: LanguageCode;
-}
