+++ src/data/signupPropositions.json
@@ -22,20 +22,20 @@
   "fr": [
     {
       "id": 1,
-      "title_fr": "FR-Fast and Easy",
-      "content_fr": "FR-Thanks to the account, you can price your insurance faster or change the insurance date",
+      "title": "FR-Fast and Easy",
+      "content": "FR-Thanks to the account, you can price your insurance faster or change the insurance date",
       "iconPath": "speedometer"
     },
     {
       "id": 2,
-      "title_fr": "FR-Recurring Activities",
-      "content_fr": "FR-Thanks to the account, you can price your insurance faster or change the insurance date",
+      "title": "FR-Recurring Activities",
+      "content": "FR-Thanks to the account, you can price your insurance faster or change the insurance date",
       "iconPath": "recurringTasks"
     },
     {
       "id": 3,
-      "title_fr": "FR-Everything in one place",
-      "content_fr": "FR-Thanks to the account, you can price your insurance faster or change the insurance date",
+      "title": "FR-Everything in one place",
+      "content": "FR-Thanks to the account, you can price your insurance faster or change the insurance date",
       "iconPath": "allInOne"
     }
   ]
